+++ javax/src/main/java/me/panpf/javax/util/Zipx.java
@@ -16,6 +16,7 @@
 
 package me.panpf.javax.util;
 
+import me.panpf.javax.io.CopyListener;
 import me.panpf.javax.io.Filex;
 import me.panpf.javax.io.IOStreamx;
 import org.jetbrains.annotations.NotNull;
@@ -33,7 +34,7 @@ import java.util.zip.ZipOutputStream;
 /**
  * ZIP tool method
  */
-@SuppressWarnings("WeakerAccess")
+@SuppressWarnings({"WeakerAccess", "UnusedReturnValue"})
 public class Zipx {
 
     private Zipx() {
@@ -45,11 +46,12 @@ public class Zipx {
      * @param sourceFiles           The file to be compressed
      * @param destinationFile       Output file
      * @param zipEntryNameTransform Get the name of the file compressed into the zip package
+     * @param listener              Progress listener
      * @return Output file
      * @throws IOException IO exceptions
      */
     public static File compressFilesTo(@Nullable File[] sourceFiles, @NotNull File destinationFile,
-                                       @NotNull Transformer<File, String> zipEntryNameTransform) throws IOException {
+                                       @NotNull Transformer<File, String> zipEntryNameTransform, @Nullable final ZipListener listener) throws IOException {
         ZipOutputStream zipOutputStream = null;
         try {
             zipOutputStream = new ZipOutputStream(new BufferedOutputStream(new FileOutputStream(destinationFile, false)));
@@ -58,6 +60,20 @@ public class Zipx {
             if (sourceFiles != null) {
                 Collectionx.addAll(files, sourceFiles);
             }
+
+            final long totalLength = listener != null ? Filex.lengthRecursively(files) : 0;
+            final long[] totalCompletedLength = new long[1];
+            final ZipEntry[] currentZipEntry = new ZipEntry[1];
+            final long[] entrySize = new long[]{-1};
+            CopyListener copyListener = new CopyListener() {
+                @Override
+                public void onUpdateProgress(long completedLength) {
+                    if (listener != null) {
+                        totalCompletedLength[0] += completedLength;
+                        listener.onUpdateProgress(totalLength, totalCompletedLength[0], currentZipEntry[0], entrySize[0], completedLength);
+                    }
+                }
+            };
 
             while (true) {
                 File childFile;
@@ -77,11 +93,16 @@ public class Zipx {
                         Collections.addAll(files, fileList);
                     }
                 } else {
-                    zipOutputStream.putNextEntry(new ZipEntry(zipEntryNameTransform.transform(childFile)));
+                    ZipEntry zipEntry = new ZipEntry(zipEntryNameTransform.transform(childFile));
+                    currentZipEntry[0] = zipEntry;
+                    entrySize[0] = childFile.length();
+                    zipOutputStream.putNextEntry(zipEntry);
                     InputStream inputStream = null;
                     try {
                         inputStream = new BufferedInputStream(new FileInputStream(childFile));
-                        IOStreamx.copyTo(inputStream, zipOutputStream);
+                        if (listener != null) listener.onEntryStart(currentZipEntry[0]);
+                        IOStreamx.copyTo(inputStream, zipOutputStream, copyListener);
+                        if (listener != null) listener.onEntryEnd(currentZipEntry[0]);
                     } finally {
                         zipOutputStream.closeEntry();
                         IOStreamx.safeClose(inputStream);
@@ -101,6 +122,63 @@ public class Zipx {
     }
 
     /**
+     * Compress the specified files
+     *
+     * @param sourceFiles           The file to be compressed
+     * @param destinationFile       Output file
+     * @param zipEntryNameTransform Get the name of the file compressed into the zip package
+     * @return Output file
+     * @throws IOException IO exceptions
+     */
+    public static File compressFilesTo(@Nullable File[] sourceFiles, @NotNull File destinationFile,
+                                       @NotNull Transformer<File, String> zipEntryNameTransform) throws IOException {
+        return compressFilesTo(sourceFiles, destinationFile, zipEntryNameTransform, null);
+    }
+
+    /**
+     * Compress the specified file
+     *
+     * @param sourceFile      The file to be compressed
+     * @param destinationFile Output file
+     * @param listener        Progress listener
+     * @return Output file
+     * @throws IOException IO exceptions
+     */
+    public static File compressFileTo(@NotNull File sourceFile, @NotNull File destinationFile, @Nullable final ZipListener listener) throws IOException {
+        return compressFilesTo(new File[]{sourceFile}, destinationFile, new Transformer<File, String>() {
+            @NotNull
+            @Override
+            public String transform(@NotNull File file) {
+                return file.getName();
+            }
+        }, listener);
+    }
+
+    /**
+     * Compress the specified file
+     *
+     * @param sourceFile      The file to be compressed
+     * @param destinationFile Output file
+     * @return Output file
+     * @throws IOException IO exceptions
+     */
+    public static File compressFileTo(@NotNull File sourceFile, @NotNull File destinationFile) throws IOException {
+        return compressFileTo(sourceFile, destinationFile, null);
+    }
+
+    /**
+     * Compress the specified file
+     *
+     * @param sourceFile The file to be compressed
+     * @param listener   Progress listener
+     * @return Output file
+     * @throws IOException IO exceptions
+     */
+    public static File compressFile(@NotNull File sourceFile, @Nullable final ZipListener listener) throws IOException {
+        return compressFileTo(sourceFile, getCompressDstFile(sourceFile), listener);
+    }
+
+    /**
      * Compress the specified file
      *
      * @param sourceFile The file to be compressed
@@ -108,13 +186,28 @@ public class Zipx {
      * @throws IOException IO exceptions
      */
     public static File compressFile(@NotNull File sourceFile) throws IOException {
-        return compressFilesTo(new File[]{sourceFile}, getCompressDstFile(sourceFile), new Transformer<File, String>() {
+        return compressFileTo(sourceFile, getCompressDstFile(sourceFile), null);
+    }
+
+    /**
+     * Compress all files in the specified directory to
+     *
+     * @param sourceDirectory The directory to be compressed
+     * @param destinationFile Output file
+     * @param listener        Progress listener
+     * @return Output file
+     * @throws IOException IO exceptions
+     */
+    public static File compressDirTo(@NotNull final File sourceDirectory, @NotNull File destinationFile, @Nullable final ZipListener listener) throws IOException {
+        Premisex.requireFileExist(sourceDirectory, "sourceDirectory");
+        Premisex.requireIsDir(sourceDirectory, "sourceDirectory");
+        return compressFilesTo(sourceDirectory.listFiles(), destinationFile, new Transformer<File, String>() {
             @NotNull
             @Override
             public String transform(@NotNull File file) {
-                return file.getName();
+                return file.getPath().replace(sourceDirectory.getPath() + File.separator, "");
             }
-        });
+        }, listener);
     }
 
     /**
@@ -126,15 +219,19 @@ public class Zipx {
      * @throws IOException IO exceptions
      */
     public static File compressDirTo(@NotNull final File sourceDirectory, @NotNull File destinationFile) throws IOException {
-        Premisex.requireFileExist(sourceDirectory, "sourceDirectory");
-        Premisex.requireIsDir(sourceDirectory, "sourceDirectory");
-        return compressFilesTo(sourceDirectory.listFiles(), destinationFile, new Transformer<File, String>() {
-            @NotNull
-            @Override
-            public String transform(@NotNull File file) {
-                return file.getPath().replace(sourceDirectory.getPath() + File.separator, "");
-            }
-        });
+        return compressDirTo(sourceDirectory, destinationFile, null);
+    }
+
+    /**
+     * Compress all files under the specified folder, output ZIP file and [sourceDirectory] in the same directory, name is [sourceDirectory] plus '.zip'
+     *
+     * @param sourceDirectory The directory to be compressed
+     * @param listener        Progress listener
+     * @return Output file
+     * @throws IOException IO exceptions
+     */
+    public static File compressDir(@NotNull File sourceDirectory, @Nullable final ZipListener listener) throws IOException {
+        return compressDirTo(sourceDirectory, getCompressDstFile(sourceDirectory), listener);
     }
 
     /**
@@ -145,7 +242,7 @@ public class Zipx {
      * @throws IOException IO exceptions
      */
     public static File compressDir(@NotNull File sourceDirectory) throws IOException {
-        return compressDirTo(sourceDirectory, getCompressDstFile(sourceDirectory));
+        return compressDir(sourceDirectory, null);
     }
 
     /**
@@ -165,7 +262,7 @@ public class Zipx {
      * @throws IOException IO exceptions. include ZipException, UnableCreateDirException, UnableCreateFileException
      */
     @NotNull
-    public static File decompressTo(@NotNull File zipSourceFile, @NotNull final File destinationDir) throws IOException {
+    public static File decompressTo(@NotNull File zipSourceFile, @NotNull final File destinationDir, @Nullable final ZipListener listener) throws IOException {
         Premisex.requireFileExist(zipSourceFile, "zipSourceFile");
         Premisex.require(!destinationDir.exists() || destinationDir.isDirectory(), new LazyValue<String>() {
             @NotNull
@@ -177,9 +274,23 @@ public class Zipx {
 
         ZipFile zipFile = new ZipFile(zipSourceFile);
         try {
-            Enumeration<?> entries = zipFile.entries();
+            final long totalLength = listener != null ? getTrueSize(zipFile) : 0;
+            final long[] totalCompletedLength = new long[1];
+            final ZipEntry[] currentZipEntry = new ZipEntry[1];
+            CopyListener copyListener = new CopyListener() {
+                @Override
+                public void onUpdateProgress(long completedLength) {
+                    if (listener != null) {
+                        totalCompletedLength[0] += completedLength;
+                        listener.onUpdateProgress(totalLength, totalCompletedLength[0], currentZipEntry[0], currentZipEntry[0].getSize(), completedLength);
+                    }
+                }
+            };
+
+            Enumeration<? extends ZipEntry> entries = zipFile.entries();
             while (entries.hasMoreElements()) {
-                ZipEntry zipEntry = (ZipEntry) entries.nextElement();
+                ZipEntry zipEntry = entries.nextElement();
+                currentZipEntry[0] = zipEntry;
                 File file = new File(destinationDir, zipEntry.getName());
                 if (zipEntry.isDirectory()) {
                     Filex.mkdirsOrThrow(file);
@@ -190,7 +301,9 @@ public class Zipx {
                     try {
                         inputStream = zipFile.getInputStream(zipEntry);
                         outputStream = new FileOutputStream(file, false);
-                        IOStreamx.copyTo(inputStream, outputStream);
+                        if (listener != null) listener.onEntryStart(currentZipEntry[0]);
+                        IOStreamx.copyTo(inputStream, outputStream, copyListener);
+                        if (listener != null) listener.onEntryEnd(currentZipEntry[0]);
                     } finally {
                         IOStreamx.safeClose(outputStream);
                         IOStreamx.safeClose(inputStream);
@@ -204,6 +317,33 @@ public class Zipx {
         return destinationDir;
     }
 
+
+    /**
+     * Decompress the ZIP file to the specified folder
+     *
+     * @param zipSourceFile  ZIP file
+     * @param destinationDir Out dir
+     * @return Out dir
+     * @throws IOException IO exceptions. include ZipException, UnableCreateDirException, UnableCreateFileException
+     */
+    @NotNull
+    public static File decompressTo(@NotNull File zipSourceFile, @NotNull final File destinationDir) throws IOException {
+        return decompressTo(zipSourceFile, destinationDir, null);
+    }
+
+    /**
+     * Decompress the ZIP file to its directory, and the output folder name is the name of the ZIP file (without the suffix)
+     *
+     * @param zipSourceFile ZIP file
+     * @param listener      Progress listener
+     * @return Out dir
+     * @throws IOException IO exceptions. include ZipException, UnableCreateDirException, UnableCreateFileException
+     */
+    @NotNull
+    public static File decompress(@NotNull File zipSourceFile, @Nullable final ZipListener listener) throws IOException {
+        return decompressTo(zipSourceFile, getDecompressDstDir(zipSourceFile), listener);
+    }
+
     /**
      * Decompress the ZIP file to its directory, and the output folder name is the name of the ZIP file (without the suffix)
      *
@@ -213,7 +353,7 @@ public class Zipx {
      */
     @NotNull
     public static File decompress(@NotNull File zipSourceFile) throws IOException {
-        return decompressTo(zipSourceFile, getDecompressDstDir(zipSourceFile));
+        return decompress(zipSourceFile, null);
     }
 
     /**
@@ -221,5 +361,27 @@ public class Zipx {
      */
     public static File getDecompressDstDir(@NotNull File zipSourceFile) {
         return new File(zipSourceFile.getParentFile(), Filex.getNameWithoutExtension(zipSourceFile));
+    }
+
+
+    /**
+     * Get the original size of the Zip file
+     */
+    public static long getTrueSize(@NotNull ZipFile zipFile) {
+        long totalSize = 0;
+        Enumeration<? extends ZipEntry> elements = zipFile.entries();
+        while (elements.hasMoreElements()) {
+            totalSize += elements.nextElement().getSize();
+        }
+        return totalSize;
+    }
+
+    /**
+     * Get the original size of the Zip file
+     */
+    public static long getTrueSize(@NotNull File file) throws IOException {
+        try (ZipFile zipFile = new ZipFile(file)) {
+            return getTrueSize(zipFile);
+        }
     }
 }